:rocket: removing unnecessary files
