+++ csharp/Facebook.CSSLayout/CSSConstants.cs
@@ -2,7 +2,7 @@
 {
     public static class CSSConstants
     {
-        public static float UNDEFINED = float.NaN;
+        public static float Undefined = float.NaN;
 
         public static bool IsUndefined(float value)
         {
