+++ csharp/Facebook.CSSLayout/Spacing.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Facebook.CSSLayout
 {
@@ -16,47 +12,47 @@ namespace Facebook.CSSLayout
         /// <summary>
         /// Spacing type that represents the left direction. E.g. <code>marginLeft</code>.
         /// </summary>
-        public const int LEFT = 0;
+        public const int Left = 0;
 
         /// <summary>
         /// Spacing type that represents the top direction. E.g. <code>marginTop</code>.
         /// </summary>
-        public const int TOP = 1;
+        public const int Top = 1;
 
         /// <summary>
         /// Spacing type that represents the right direction. E.g. <code>marginRight</code>.
         /// </summary>
-        public const int RIGHT = 2;
+        public const int Right = 2;
 
         /// <summary>
         /// Spacing type that represents the bottom direction. E.g. <code>marginBottom</code>.
         /// </summary>
-        public const int BOTTOM = 3;
+        public const int Bottom = 3;
 
         /// <summary>
         /// Spacing type that represents start direction e.g. left in left-to-right, right in right-to-left.
         /// </summary>
-        public const int START = 4;
+        public const int Start = 4;
 
         /// <summary>
         /// Spacing type that represents end direction e.g. right in left-to-right, left in right-to-left.
         /// </summary>
-        public const int END = 5;
+        public const int End = 5;
 
         /// <summary>
         /// Spacing type that represents horizontal direction (left and right). E.g. <code>marginHorizontal</code>.
         /// </summary>
-        public const int HORIZONTAL = 6;
+        public const int Horizontal = 6;
 
         /// <summary>
         /// Spacing type that represents vertical direction (top and bottom). E.g. <code>marginVertical</code>.
         /// </summary>
-        public const int VERTICAL = 7;
+        public const int Vertical = 7;
 
         /// <summary>
         /// Spacing type that represents all directions (left, top, right, bottom). E.g. <code>margin</code>.
         /// </summary>
-        public const int ALL = 8;
+        public const int All = 8;
 
         private static readonly int[] sFlagsMap = new[]
         {
@@ -71,10 +67,10 @@ namespace Facebook.CSSLayout
             256, /*ALL*/
         };
 
-        private int mValueFlags = 0;
-        private readonly float mDefaultValue;
-        private readonly float[] mSpacing = NewFullSpacingArray();
-        private bool mHasAliasesSet;
+        private int _valueFlags = 0;
+        private readonly float _defaultValue;
+        private readonly float[] _spacing = NewFullSpacingArray();
+        private bool _hasAliasesSet;
 
         public Spacing() : this(0)
         {
@@ -83,33 +79,33 @@ namespace Facebook.CSSLayout
 
         public Spacing(float defaultValue)
         {
-            mDefaultValue = defaultValue;
+            _defaultValue = defaultValue;
         }
 
         public float Get(int spacingType)
         {
-            var defaultValue = spacingType == START || spacingType == END ? CSSConstants.UNDEFINED : mDefaultValue;
+            var defaultValue = spacingType == Start || spacingType == End ? CSSConstants.Undefined : _defaultValue;
 
-            if (mValueFlags == 0)
+            if (_valueFlags == 0)
             {
                 return defaultValue;
             }
 
-            if ((mValueFlags & sFlagsMap[spacingType]) != 0)
+            if ((_valueFlags & sFlagsMap[spacingType]) != 0)
             {
-                return mSpacing[spacingType];
+                return _spacing[spacingType];
             }
 
-            if (mHasAliasesSet)
+            if (_hasAliasesSet)
             {
-                var secondType = spacingType == TOP || spacingType == BOTTOM ? VERTICAL : HORIZONTAL;
-                if ((mValueFlags & sFlagsMap[secondType]) != 0)
+                var secondType = spacingType == Top || spacingType == Bottom ? Vertical : Horizontal;
+                if ((_valueFlags & sFlagsMap[secondType]) != 0)
                 {
-                    return mSpacing[secondType];
+                    return _spacing[secondType];
                 }
-                else if ((mValueFlags & sFlagsMap[ALL]) != 0)
+                else if ((_valueFlags & sFlagsMap[All]) != 0)
                 {
-                    return mSpacing[ALL];
+                    return _spacing[All];
                 }
             }
 
@@ -118,28 +114,28 @@ namespace Facebook.CSSLayout
 
         public float GetRaw(int spacingType)
         {
-            return mSpacing[spacingType];
+            return _spacing[spacingType];
         }
 
         public bool Set(int spacingType, float value)
         {
-            if (FloatEqual(mSpacing[spacingType], value))
+            if (FloatEqual(_spacing[spacingType], value))
             {
-                mSpacing[spacingType] = value;
+                _spacing[spacingType] = value;
 
                 if (CSSConstants.IsUndefined(value))
                 {
-                    mValueFlags &= ~sFlagsMap[spacingType];
+                    _valueFlags &= ~sFlagsMap[spacingType];
                 }
                 else
                 {
-                    mValueFlags |= ~sFlagsMap[spacingType];
+                    _valueFlags |= ~sFlagsMap[spacingType];
                 }
 
-                mHasAliasesSet =
-                    (mValueFlags & sFlagsMap[ALL]) != 0 ||
-                    (mValueFlags & sFlagsMap[VERTICAL]) != 0 ||
-                    (mValueFlags & sFlagsMap[HORIZONTAL]) != 0;
+                _hasAliasesSet =
+                    (_valueFlags & sFlagsMap[All]) != 0 ||
+                    (_valueFlags & sFlagsMap[Vertical]) != 0 ||
+                    (_valueFlags & sFlagsMap[Horizontal]) != 0;
 
                 return true;
             }
@@ -149,9 +145,9 @@ namespace Facebook.CSSLayout
 
         public void Reset()
         {
-            for (var i = 0; i < 9; i++) { mSpacing[i] = CSSConstants.UNDEFINED; }
-            mHasAliasesSet = false;
-            mValueFlags = 0;
+            for (var i = 0; i < 9; i++) { _spacing[i] = CSSConstants.Undefined; }
+            _hasAliasesSet = false;
+            _valueFlags = 0;
         }
 
         private static bool FloatEqual(float f1, float f2)
@@ -168,15 +164,15 @@ namespace Facebook.CSSLayout
         {
             return new float[]
             {
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED,
-                CSSConstants.UNDEFINED
+                CSSConstants.Undefined,
+                CSSConstants.Undefined,
+                CSSConstants.Undefined,
+                CSSConstants.Undefined,
+                CSSConstants.Undefined,
+                CSSConstants.Undefined,
+                CSSConstants.Undefined,
+                CSSConstants.Undefined,
+                CSSConstants.Undefined
             };
         }
     }
